Did some refactoring
