+++ jctool/jctool_ui.hpp
@@ -1,0 +1,9 @@
+#pragma once
+#include "Controller.hpp"
+
+namespace JCToolkit {
+    namespace UI {
+        void show(Controller& controller, RumbleData& rumble_data);
+    }
+    void init();
+}
